Fix STATIC_ASSERT on gcc and change string sizes to 32bit ints.

+++ Include/Rocket/Core/StringBase.h
@@ -41,7 +41,7 @@ template< typename T >
 class StringBase
 {
 public:
-	typedef size_t size_type;
+	typedef unsigned int size_type;
 	static const size_type npos = (size_type)-1;
 
 	StringBase();
